+++ docs/scripts/buildApi.ts
@@ -1,41 +1,17 @@
-import { mkdirSync, readFileSync, writeFileSync } from 'fs';
+import { mkdirSync, readFileSync } from 'fs';
 import * as fse from 'fs-extra';
 import path from 'path';
-import * as babel from '@babel/core';
-import traverse from '@babel/traverse';
 import * as _ from 'lodash';
 import kebabCase from 'lodash/kebabCase';
-import * as prettier from 'prettier';
-import remark from 'remark';
-import remarkVisit from 'unist-util-visit';
 import * as yargs from 'yargs';
-import { defaultHandlers, parse as docgenParse, ReactDocgenApi } from 'react-docgen';
-import muiDefaultPropsHandler from 'docs/src/modules/utils/defaultPropsHandler';
-import { LANGUAGES } from 'docs/src/modules/constants';
-import parseTest from 'docs/src/modules/utils/parseTest';
-import generatePropTypeDescription, {
-  getChained,
-} from 'docs/src/modules/utils/generatePropTypeDescription';
+import { ReactDocgenApi } from 'react-docgen';
 import { findPages, findPagesMarkdown, findComponents } from 'docs/src/modules/utils/find';
-import { getHeaders, renderInline as renderMarkdownInline } from '@mui/markdown';
-import { pageToTitle } from 'docs/src/modules/utils/helpers';
-import createDescribeableProp, {
-  DescribeablePropDescriptor,
-} from 'docs/src/modules/utils/createDescribeableProp';
-import generatePropDescription from 'docs/src/modules/utils/generatePropDescription';
-import parseStyles, { Styles } from 'docs/src/modules/utils/parseStyles';
-import generateUtilityClass from '@mui/base/generateUtilityClass';
+import { getHeaders } from '@mui/markdown';
+import { Styles } from 'docs/src/modules/utils/parseStyles';
 import * as ttp from 'typescript-to-proptypes';
-import { getLineFeed, getUnstyledFilename } from './helpers';
-import {
-  findComponentDemos,
-  getMuiName,
-  getGeneralPathInfo,
-  getMaterialPathInfo,
-  getBasePathInfo,
-} from './buildApiUtils';
+import { getGeneralPathInfo, getMaterialPathInfo, getBasePathInfo } from './buildApiUtils';
+import buildComponentApi, { writePrettifiedFile } from './ApiBuilders/ComponentApiBuilder';
 
-const DEFAULT_PRETTIER_CONFIG_PATH = path.join(process.cwd(), 'prettier.config.js');
 const apiDocsTranslationsDirectory = path.resolve('docs', 'translations', 'api-docs');
 
 interface ReactApi extends ReactDocgenApi {
@@ -74,454 +50,6 @@ interface ReactApi extends ReactDocgenApi {
     classDescriptions: { [key: string]: { description: string; conditions?: string } };
   };
 }
-
-const cssComponents = ['Box', 'Grid', 'Typography', 'Stack'];
-
-function writePrettifiedFile(
-  filename: string,
-  data: string,
-  prettierConfigPath: string = DEFAULT_PRETTIER_CONFIG_PATH,
-  options: object = {},
-) {
-  const prettierConfig = prettier.resolveConfig.sync(filename, {
-    config: prettierConfigPath,
-  });
-  if (prettierConfig === null) {
-    throw new Error(
-      `Could not resolve config for '${filename}' using prettier config path '${prettierConfigPath}'.`,
-    );
-  }
-
-  writeFileSync(filename, prettier.format(data, { ...prettierConfig, filepath: filename }), {
-    encoding: 'utf8',
-    ...options,
-  });
-}
-
-/**
- * Produces markdown of the description that can be hosted anywhere.
- *
- * By default we assume that the markdown is hosted on mui.com which is
- * why the source includes relative url. We transform them to absolute urls with
- * this method.
- */
-async function computeApiDescription(api: ReactApi, options: { host: string }): Promise<string> {
-  const { host } = options;
-  const file = await remark()
-    .use(function docsLinksAttacher() {
-      return function transformer(tree) {
-        remarkVisit(tree, 'link', (linkNode) => {
-          if ((linkNode.url as string).startsWith('/')) {
-            linkNode.url = `${host}${linkNode.url}`;
-          }
-        });
-      };
-    })
-    .process(api.description);
-
-  return file.contents.toString('utf-8').trim();
-}
-
-/**
- * Add demos & API comment block to type definitions, e.g.:
- * /**
- *  * Demos:
- *  *
- *  * - [Icons](https://mui.com/components/icons/)
- *  * - [Material Icons](https://mui.com/components/material-icons/)
- *  *
- *  * API:
- *  *
- *  * - [Icon API](https://mui.com/api/icon/)
- */
-async function annotateComponentDefinition(api: ReactApi) {
-  const HOST = 'https://mui.com';
-
-  const typesFilename = api.filename.replace(/\.js$/, '.d.ts');
-  const typesSource = readFileSync(typesFilename, { encoding: 'utf8' });
-  const typesAST = await babel.parseAsync(typesSource, {
-    configFile: false,
-    filename: typesFilename,
-    presets: [require.resolve('@babel/preset-typescript')],
-  });
-  if (typesAST === null) {
-    throw new Error('No AST returned from babel.');
-  }
-
-  let start = 0;
-  let end = null;
-  traverse(typesAST, {
-    ExportDefaultDeclaration(babelPath) {
-      /**
-       * export default function Menu() {}
-       */
-      let node: babel.Node = babelPath.node;
-      if (node.declaration.type === 'Identifier') {
-        // declare const Menu: {};
-        // export default Menu;
-        if (babel.types.isIdentifier(babelPath.node.declaration)) {
-          const bindingId = babelPath.node.declaration.name;
-          const binding = babelPath.scope.bindings[bindingId];
-
-          // The JSDoc MUST be located at the declaration
-          if (babel.types.isFunctionDeclaration(binding.path.node)) {
-            // For function declarations the binding is equal to the declaration
-            // /**
-            //  */
-            // function Component() {}
-            node = binding.path.node;
-          } else {
-            // For variable declarations the binding points to the declarator.
-            // /**
-            //  */
-            // const Component = () => {}
-            node = binding.path.parentPath.node;
-          }
-        }
-      }
-
-      const { leadingComments } = node;
-      const leadingCommentBlocks =
-        leadingComments != null
-          ? leadingComments.filter(({ type }) => type === 'CommentBlock')
-          : null;
-      const jsdocBlock = leadingCommentBlocks != null ? leadingCommentBlocks[0] : null;
-      if (leadingCommentBlocks != null && leadingCommentBlocks.length > 1) {
-        throw new Error(
-          `Should only have a single leading jsdoc block but got ${
-            leadingCommentBlocks.length
-          }:\n${leadingCommentBlocks
-            .map(({ type, value }, index) => `#${index} (${type}): ${value}`)
-            .join('\n')}`,
-        );
-      }
-      if (jsdocBlock != null) {
-        start = jsdocBlock.start;
-        end = jsdocBlock.end;
-      } else if (node.start !== null) {
-        start = node.start - 1;
-        end = start;
-      }
-    },
-  });
-
-  if (end === null || start === 0) {
-    throw new TypeError(
-      "Don't know where to insert the jsdoc block. Probably no `default export` found",
-    );
-  }
-
-  let inheritanceAPILink = null;
-  if (api.inheritance !== null) {
-    const url = api.inheritance.pathname.startsWith('/')
-      ? `${HOST}${api.inheritance.pathname}`
-      : api.inheritance.pathname;
-
-    inheritanceAPILink = `[${api.inheritance.component} API](${url})`;
-  }
-
-  const markdownLines = (await computeApiDescription(api, { host: HOST })).split('\n');
-  // Ensure a newline between manual and generated description.
-  if (markdownLines[markdownLines.length - 1] !== '') {
-    markdownLines.push('');
-  }
-  markdownLines.push(
-    'Demos:',
-    '',
-    ...api.demos.map((demoPathname) => {
-      return `- [${pageToTitle({ pathname: demoPathname })}](${HOST}${demoPathname}/)`;
-    }),
-    '',
-  );
-
-  markdownLines.push('API:', '', `- [${api.name} API](${HOST}${api.apiUrl}/)`);
-  if (api.inheritance !== null) {
-    markdownLines.push(`- inherits ${inheritanceAPILink}`);
-  }
-
-  const jsdoc = `/**\n${markdownLines
-    .map((line) => (line.length > 0 ? ` * ${line}` : ` *`))
-    .join('\n')}\n */`;
-  const typesSourceNew = typesSource.slice(0, start) + jsdoc + typesSource.slice(end);
-  writeFileSync(typesFilename, typesSourceNew, { encoding: 'utf8' });
-}
-
-/**
- * Substitute CSS class description conditions with placeholder
- */
-function extractClassConditions(descriptions: any) {
-  const classConditions: {
-    [key: string]: { description: string; conditions?: string; nodeName?: string };
-  } = {};
-  const stylesRegex =
-    /((Styles|State class|Class name) applied to )(.*?)(( if | unless | when |, ){1}(.*))?\./;
-
-  Object.entries(descriptions).forEach(([className, description]: any) => {
-    if (className) {
-      const conditions = description.match(stylesRegex);
-
-      if (conditions && conditions[6]) {
-        classConditions[className] = {
-          description: description.replace(stylesRegex, '$1{{nodeName}}$5{{conditions}}.'),
-          nodeName: conditions[3],
-          conditions: conditions[6].replace(/`(.*?)`/g, '<code>$1</code>'),
-        };
-      } else if (conditions && conditions[3] && conditions[3] !== 'the root element') {
-        classConditions[className] = {
-          description: description.replace(stylesRegex, '$1{{nodeName}}$5.'),
-          nodeName: conditions[3],
-        };
-      } else {
-        classConditions[className] = { description };
-      }
-    }
-  });
-  return classConditions;
-}
-
-/**
- * Generate list of component demos
- */
-function generateDemoList(reactAPI: ReactApi): string {
-  return `<ul>${reactAPI.demos
-    .map(
-      (demoPathname) =>
-        `<li><a href="${demoPathname}/">${pageToTitle({ pathname: demoPathname })}</a></li>`,
-    )
-    .join('\n')}</ul>`;
-}
-
-/**
- * @param filepath - absolute path
- * @example toGithubPath('/home/user/material-ui/packages/Accordion') === '/packages/Accordion'
- * @example toGithubPath('C:\\Development\material-ui\packages\Accordion') === '/packages/Accordion'
- */
-function toGithubPath(filepath: string): string {
-  return `/${path.relative(process.cwd(), filepath).replace(/\\/g, '/')}`;
-}
-
-const generateApiTranslations = (outputDirectory: string, reactApi: ReactApi) => {
-  const componentName = reactApi.name;
-  const apiDocsTranslationPath = path.resolve(outputDirectory, kebabCase(componentName));
-  function resolveApiDocsTranslationsComponentLanguagePath(language: typeof LANGUAGES[0]): string {
-    const languageSuffix = language === 'en' ? '' : `-${language}`;
-
-    return path.join(apiDocsTranslationPath, `${kebabCase(componentName)}${languageSuffix}.json`);
-  }
-
-  mkdirSync(apiDocsTranslationPath, {
-    mode: 0o777,
-    recursive: true,
-  });
-
-  writePrettifiedFile(
-    resolveApiDocsTranslationsComponentLanguagePath('en'),
-    JSON.stringify(reactApi.translations),
-  );
-
-  LANGUAGES.forEach((language) => {
-    if (language !== 'en') {
-      try {
-        writePrettifiedFile(
-          resolveApiDocsTranslationsComponentLanguagePath(language),
-          JSON.stringify(reactApi.translations),
-          undefined,
-          { flag: 'wx' },
-        );
-      } catch (error) {
-        // File exists
-      }
-    }
-  });
-};
-
-const generateApiPage = (outputDirectory: string, reactApi: ReactApi) => {
-  /**
-   * Gather the metadata needed for the component's API page.
-   */
-  const pageContent = {
-    // Sorted by required DESC, name ASC
-    props: _.fromPairs(
-      Object.entries(reactApi.propsTable).sort(([aName, aData], [bName, bData]) => {
-        if ((aData.required && bData.required) || (!aData.required && !bData.required)) {
-          return aName.localeCompare(bName);
-        }
-        if (aData.required) {
-          return -1;
-        }
-        return 1;
-      }),
-    ),
-    name: reactApi.name,
-    styles: {
-      classes: reactApi.styles.classes,
-      globalClasses: _.fromPairs(
-        Object.entries(reactApi.styles.globalClasses).filter(([className, globalClassName]) => {
-          // Only keep "non-standard" global classnames
-          return globalClassName !== `Mui${reactApi.name}-${className}`;
-        }),
-      ),
-      name: reactApi.styles.name,
-    },
-    spread: reactApi.spread,
-    forwardsRefTo: reactApi.forwardsRefTo,
-    filename: toGithubPath(reactApi.filename),
-    inheritance: reactApi.inheritance,
-    demos: generateDemoList(reactApi),
-    cssComponent: cssComponents.indexOf(reactApi.name) >= 0,
-  };
-
-  writePrettifiedFile(
-    path.resolve(outputDirectory, `${kebabCase(reactApi.name)}.json`),
-    JSON.stringify(pageContent),
-  );
-
-  writePrettifiedFile(
-    path.resolve(outputDirectory, `${kebabCase(reactApi.name)}.js`),
-    `import * as React from 'react';
-import ApiPage from 'docs/src/modules/components/ApiPage';
-import mapApiPageTranslations from 'docs/src/modules/utils/mapApiPageTranslations';
-import jsonPageContent from './${kebabCase(reactApi.name)}.json';
-
-export default function Page(props) {
-  const { descriptions, pageContent } = props;
-  return <ApiPage descriptions={descriptions} pageContent={pageContent} />;
-}
-
-Page.getInitialProps = () => {
-  const req = require.context(
-    'docs/translations/api-docs/${kebabCase(reactApi.name)}',
-    false,
-    /${kebabCase(reactApi.name)}.*.json$/,
-  );
-  const descriptions = mapApiPageTranslations(req);
-
-  return {
-    descriptions,
-    pageContent: jsonPageContent,
-  };
-};
-`.replace(/\r?\n/g, reactApi.EOL),
-  );
-};
-
-const attachTranslations = (reactApi: ReactApi) => {
-  const translations: ReactApi['translations'] = {
-    componentDescription: reactApi.description,
-    propDescriptions: {},
-    classDescriptions: {},
-  };
-  Object.entries(reactApi.props!).forEach(([propName, propDescriptor]) => {
-    let prop: DescribeablePropDescriptor | null;
-    try {
-      prop = createDescribeableProp(propDescriptor, propName);
-    } catch (error) {
-      prop = null;
-    }
-    if (prop) {
-      let description = generatePropDescription(prop, propName);
-      description = renderMarkdownInline(description);
-
-      if (propName === 'classes') {
-        description += ' See <a href="#css">CSS API</a> below for more details.';
-      } else if (propName === 'sx') {
-        description += ' See the <a href="/system/the-sx-prop/">`sx` page</a> for more details.';
-      }
-      translations.propDescriptions[propName] = description.replace(/\n@default.*$/, '');
-    }
-  });
-
-  /**
-   * CSS class descriptiohs.
-   */
-  translations.classDescriptions = extractClassConditions(reactApi.styles.descriptions);
-
-  reactApi.translations = translations;
-};
-
-const attachPropsTable = (reactApi: ReactApi) => {
-  const propErrors: Array<[propName: string, error: Error]> = [];
-  const componentProps: ReactApi['propsTable'] = _.fromPairs(
-    Object.entries(reactApi.props!).map(([propName, propDescriptor]) => {
-      let prop: DescribeablePropDescriptor | null;
-      try {
-        prop = createDescribeableProp(propDescriptor, propName);
-      } catch (error) {
-        propErrors.push([propName, error as Error]);
-        prop = null;
-      }
-      if (prop === null) {
-        // have to delete `componentProps.undefined` later
-        return [] as any;
-      }
-
-      // Only keep `default` for bool props if it isn't 'false'.
-      let defaultValue: string | undefined;
-      if (
-        propDescriptor.type.name !== 'bool' ||
-        propDescriptor.jsdocDefaultValue?.value !== 'false'
-      ) {
-        defaultValue = propDescriptor.jsdocDefaultValue?.value;
-      }
-
-      const propTypeDescription = generatePropTypeDescription(propDescriptor.type);
-      const chainedPropType = getChained(prop.type);
-
-      const requiredProp =
-        prop.required ||
-        /\.isRequired/.test(prop.type.raw) ||
-        (chainedPropType !== false && chainedPropType.required);
-
-      const deprecation = (propDescriptor.description || '').match(/@deprecated(\s+(?<info>.*))?/);
-
-      return [
-        propName,
-        {
-          type: {
-            name: propDescriptor.type.name,
-            description:
-              propTypeDescription !== propDescriptor.type.name ? propTypeDescription : undefined,
-          },
-          default: defaultValue,
-          // undefined values are not serialized => saving some bytes
-          required: requiredProp || undefined,
-          deprecated: !!deprecation || undefined,
-          deprecationInfo:
-            renderMarkdownInline(deprecation?.groups?.info || '').trim() || undefined,
-        },
-      ];
-    }),
-  );
-  if (propErrors.length > 0) {
-    throw new Error(
-      `There were errors creating prop descriptions:\n${propErrors
-        .map(([propName, error]) => {
-          return `  - ${propName}: ${error}`;
-        })
-        .join('\n')}`,
-    );
-  }
-
-  // created by returning the `[]` entry
-  delete componentProps.undefined;
-
-  reactApi.propsTable = componentProps;
-};
-
-const parseFile = (filename: string) => {
-  const src = readFileSync(filename, 'utf8');
-  return {
-    src,
-    shouldSkip:
-      filename.indexOf('internal') !== -1 ||
-      !!src.match(/@ignore - internal component\./) ||
-      !!src.match(/@ignore - do not document\./),
-    spread: !src.match(/ = exactProp\(/),
-    name: path.parse(filename).name,
-    EOL: getLineFeed(src),
-    inheritedComponent: src.match(/\/\/ @inheritedComponent (.*)/)?.[1],
-  };
-};
 
 async function removeOutdatedApiDocsTranslations(components: readonly ReactApi[]): Promise<void> {
   const componentDirectories = new Set<string>();
@@ -576,10 +104,6 @@ interface Settings {
      */
     pagesDirectory: string;
     /**
-     * translations output directory
-     */
-    apiTranslationsDirectory: string;
-    /**
      * The output path of `pagesApi` generated from `input.pageDirectory`
      */
     apiManifestPath: string;
@@ -604,7 +128,6 @@ const BEFORE_MIGRATION_SETTINGS: Settings[] = [
     },
     output: {
       pagesDirectory: path.join(process.cwd(), 'docs/pages/api-docs'),
-      apiTranslationsDirectory: path.join(process.cwd(), 'docs/translations/api-docs'),
       apiManifestPath: path.join(process.cwd(), 'docs/src/pagesApi.js'),
     },
     productUrlPrefix: '',
@@ -632,7 +155,6 @@ const MIGRATION_SETTINGS: Settings[] = [
     },
     output: {
       pagesDirectory: path.join(process.cwd(), 'docs/pages/material/api-docs'),
-      apiTranslationsDirectory: path.join(process.cwd(), 'docs/translations/api-docs'),
       apiManifestPath: path.join(process.cwd(), 'docs/data/material/pagesApi.js'),
     },
     productUrlPrefix: '/material',
@@ -658,7 +180,6 @@ const POST_MIGRATION_SETTINGS: Settings[] = [
     },
     output: {
       pagesDirectory: path.join(process.cwd(), 'docs/pages/material/api-docs'),
-      apiTranslationsDirectory: path.join(process.cwd(), 'docs/translations/api-docs'),
       apiManifestPath: path.join(process.cwd(), 'docs/data/material/pagesApi.js'),
     },
     productUrlPrefix: '/material',
@@ -672,12 +193,12 @@ const POST_MIGRATION_SETTINGS: Settings[] = [
     },
     output: {
       pagesDirectory: path.join(process.cwd(), 'docs/pages/base/api-docs'),
-      apiTranslationsDirectory: path.join(process.cwd(), 'docs/translations/api-docs'),
       apiManifestPath: path.join(process.cwd(), 'docs/data/base/pagesApi.js'),
     },
     productUrlPrefix: '/base',
     getPathInfo: getBasePathInfo,
   },
+  // add other products, eg. joy, data-grid, ...etc
 ];
 
 const ACTIVE_SETTINGS = BEFORE_MIGRATION_SETTINGS;
@@ -693,7 +214,6 @@ async function run(argv: { grep?: string }) {
     const componentDirectories = setting.input.libDirectory;
     const apiPagesManifestPath = setting.output.apiManifestPath;
     const pagesDirectory = setting.output.pagesDirectory;
-    const apiTranslationsDirectory = setting.output.apiTranslationsDirectory;
 
     mkdirSync(pagesDirectory, { mode: 0o777, recursive: true });
     const manifestDir = apiPagesManifestPath.match(/(.*)\/[^/]+\./)?.[1];
@@ -761,122 +281,14 @@ async function run(argv: { grep?: string }) {
       try {
         const { filename } = component;
         const pathInfo = setting.getPathInfo(filename);
-        const { shouldSkip, name, spread, EOL, inheritedComponent } = parseFile(filename);
 
-        if (shouldSkip) {
-          return null;
-        }
-
-        const reactApi: ReactApi = docgenParse(
-          readFileSync(filename, 'utf8'),
-          null,
-          defaultHandlers.concat(muiDefaultPropsHandler),
-          { filename },
-        );
-
-        // === Handle unstyled component ===
-        const unstyledFileName = getUnstyledFilename(filename);
-        let unstyledSrc;
-
-        // Try to get data for the unstyled component
-        try {
-          unstyledSrc = readFileSync(unstyledFileName, 'utf8');
-        } catch (err) {
-          // Unstyled component does not exist
-        }
-
-        if (unstyledSrc) {
-          const unstyledReactAPI = docgenParse(
-            unstyledSrc,
-            null,
-            defaultHandlers.concat(muiDefaultPropsHandler),
-            {
-              filename: unstyledFileName,
-            },
-          );
-
-          Object.keys(unstyledReactAPI.props).forEach((prop) => {
-            if (
-              unstyledReactAPI.props[prop].defaultValue &&
-              reactApi.props &&
-              (!reactApi.props[prop] || !reactApi.props[prop].defaultValue)
-            ) {
-              if (reactApi.props[prop]) {
-                reactApi.props[prop].defaultValue = unstyledReactAPI.props[prop].defaultValue;
-                reactApi.props[prop].jsdocDefaultValue =
-                  unstyledReactAPI.props[prop].jsdocDefaultValue;
-              } else {
-                reactApi.props[prop] = unstyledReactAPI.props[prop];
-              }
-            }
-          });
-        } // ================================
-
-        // Ignore what we might have generated in `annotateComponentDefinition`
-        const annotatedDescriptionMatch = reactApi.description.match(/(Demos|API):\r?\n\r?\n/);
-        if (annotatedDescriptionMatch !== null) {
-          reactApi.description = reactApi.description
-            .slice(0, annotatedDescriptionMatch.index)
-            .trim();
-        }
-        reactApi.filename = filename;
-        reactApi.name = name;
-        reactApi.apiUrl = pathInfo.apiUrl;
-        reactApi.EOL = EOL;
-        reactApi.demos = findComponentDemos(name, pagesMarkdown);
-        if (reactApi.demos.length === 0) {
-          throw new Error(
-            'Unable to find demos. \n' +
-              `Be sure to include \`components: ${reactApi.name}\` in the markdown pages where the \`${reactApi.name}\` component is relevant. ` +
-              'Every public component should have a demo. ',
-          );
-        }
-
-        const testInfo = await parseTest(filename);
-        // no Object.assign to visually check for collisions
-        reactApi.forwardsRefTo = testInfo.forwardsRefTo;
-        reactApi.spread = testInfo.spread ?? spread;
-
-        const inheritedComponentName = testInfo.inheritComponent || inheritedComponent;
-        if (inheritedComponentName) {
-          reactApi.inheritance = {
-            component: inheritedComponentName,
-            pathname:
-              inheritedComponentName === 'Transition'
-                ? 'http://reactcommunity.org/react-transition-group/transition/#Transition-props'
-                : `${setting.productUrlPrefix}/api/${kebabCase(inheritedComponentName)}/`,
-          };
-        } else {
-          reactApi.inheritance = null;
-        }
-
-        reactApi.styles = await parseStyles(reactApi, program);
-
-        if (reactApi.styles.classes.length > 0 && !reactApi.name.endsWith('Unstyled')) {
-          reactApi.styles.name = getMuiName(reactApi.name);
-        }
-        reactApi.styles.classes.forEach((key) => {
-          const globalClass = generateUtilityClass(
-            reactApi.styles.name || getMuiName(reactApi.name),
-            key,
-          );
-          reactApi.styles.globalClasses[key] = globalClass;
+        return buildComponentApi(filename, {
+          ttpProgram: program,
+          pagesMarkdown,
+          apiUrl: pathInfo.apiUrl,
+          productUrlPrefix: setting.productUrlPrefix,
+          outputPagesDirectory: setting.output.pagesDirectory,
         });
-
-        attachPropsTable(reactApi);
-
-        attachTranslations(reactApi);
-
-        generateApiTranslations(apiTranslationsDirectory, reactApi);
-
-        generateApiPage(pagesDirectory, reactApi);
-
-        // eslint-disable-next-line no-console
-        console.log('Built API docs for', reactApi.name);
-
-        await annotateComponentDefinition(reactApi);
-
-        return reactApi;
       } catch (error: any) {
         error.message = `${path.relative(process.cwd(), component.filename)}: ${error.message}`;
         throw error;
