+++ docs/scripts/tsconfig.json
@@ -1,6 +1,6 @@
 {
   "extends": "../../tsconfig.json",
-  "include": ["../types/react-docgen.d.ts", "buildApi.ts", "i18n.js"],
+  "include": ["../types/react-docgen.d.ts", "buildApi.ts", "i18n.js", "ApiBuilders/**/*"],
   "compilerOptions": {
     "allowJs": true,
     "isolatedModules": true,
